Make src a package and move config into src
This was done to make importing and installing more straightforward.
