--- config/definitions.py
@@ -1,4 +1,0 @@
-import os
-from pathlib import Path
-
-ROOT_DIR = Path(os.path.abspath(__file__)).parents[1]
