+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='src',
+    packages=find_packages(),
+    version='0.1.0'
+)
