+++ src/config/definitions.py
@@ -1,0 +1,4 @@
+import os
+from pathlib import Path
+
+ROOT_DIR = Path(os.path.abspath(__file__)).parents[2]
